+++ src/app/fcomponents/admin/admin-cart/admin-cart.component.html
@@ -1,3 +1,73 @@
-<p>
-  admin-cart works!
-</p>
+<div class="text-center mt-5 pt-5" *ngIf="isLoading || feedbackMessage">
+  <app-loading [feedbackMessage]= "feedbackMessage"></app-loading>
+</div>
+
+<div>
+  <div class="mt-3">
+    <h2>Carts</h2>
+  </div>
+  <table class="table">
+    <thead>
+      <tr class="table-secondary">
+        <th scope="col">Id<br><span>编号</span></th>
+        <th scope="col">Detail<br><span>详情</span></th>
+        <th scope="col">Contact<br><span>顾客信息</span></th>
+        <th scope="col">Price<br><span>价格</span></th>
+        <th scope="col">Location<br><span>活动地点</span></th>
+        <th scope="col">CreateTime<br><span>创建时间</span></th>
+        <th scope="col">EventDate<br><span>活动日期</span></th>
+        <th scope="col">UpdateTime<br><span>修改时间</span></th>
+        <th scope="col">Delete<br><span>删除</span></th>
+      </tr>
+    </thead>
+    <tbody *ngIf="displayedCartData">
+      <tr *ngFor="let data of displayedCartData">
+        <td>{{data.cartId}}</td>
+        <!--CartProds-->
+        <td>
+          <div class="dropdown">
+            <button class="btn btn-secondary" type="button" (click)="openCart(data)">
+              View Detail
+            </button>
+            <!-- <div class="dropdown-menu" aria-labelledby="dropdownMenuButton" *ngIf="data.cartProd">
+              <div *ngFor="let cart of data.cartProd">
+                <div class="dropdown-item">
+                  <b>{{cart.title}}</b> |
+                  <b>${{cart.price}}</b> |
+                  Qty:<b>{{cart.quantity}}</b>
+                </div>
+              </div>
+              <a class="dropdown-item" *ngIf="data.cartProd.length==0">Empty</a>
+            </div> -->
+          </div>
+        </td>
+        <!--End CartProds-->
+        <!--Contact-->
+        <td>
+          <div class="dropdown">
+            <button class="btn btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton"
+              data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+              View Contact
+            </button>
+            <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
+              <div class="dropdown-item" *ngIf="data.contact">
+                Firstname: <b>{{data.contact.firstName}}</b><br>
+                Lastname: <b>{{data.contact.lastName}}</b><br>
+                Mob: <b>{{data.contact.phoneNum}}</b><br>
+                Email: <b>{{data.contact.email}}</b>
+              </div>
+            </div>
+          </div>
+        </td>
+        <!--End Contact-->
+        <td>${{data.price}}</td>
+        <td>{{data.location}}</td>
+        <td style="font-size:10px;">{{data.createOn | date:'mediumDate'}}</td>
+        <td style="font-size:10px;">{{data.plannedTime | date:'short'}}</td>
+        <td style="font-size:10px;">{{data.updateOn | date:'mediumDate'}}</td>
+        <!-- <td><i class="far fa-edit" (click)="openCart(data)" title="Edit"></i></td> -->
+        <td ><i class="far fa-trash-alt" title="Delete" (click)="deleteCart(data)"></i></td>
+      </tr>
+    </tbody>
+  </table>
+</div>
