+++ src/app/fcomponents/admin/admin.component.css
@@ -28,24 +28,3 @@ td{
 .table td{
     padding: 0.4rem;
 }
-
-.eventTypeImage {
-    height: 100px;
-    width: 100px;
-}
-.eventTypeImageEditor{
-    display: flex;
-    justify-content: space-between;
-    align-items: flex-start;
-    flex-wrap: wrap;
-}
-.eventTypeImageEditorBlocks{
-    width: 50%;
-}
-.btn.eventTypeImageToggle:focus,.btn.eventTypeImageToggle:active {
-    outline: none !important;
-    box-shadow: none;
- }
- .card.eventTypeImageToggle {
-     border:0px;
- }
