+++ src/app/fcomponents/admin/admin.component.html
@@ -58,11 +58,9 @@
         <router-outlet></router-outlet>
 
       <!-- loading -->
-      <div class="text-center mt-5 pt-5" *ngIf="isLoading">
+      <!-- <div class="text-center mt-5 pt-5" *ngIf="isLoading">
         <app-loading></app-loading>
-      </div>
-
-      <div *ngIf="!isLoading">
+      </div> -->
 
       <!-- !Dashboard -->
       <!-- <div *ngIf="blockCode == '1'" style="text-align:center;padding-top:4%;">
@@ -221,7 +219,7 @@
       </div> -->
 
       <!-- !Faq -->
-      <div *ngIf="blockCode == '6'">
+      <!-- <div *ngIf="blockCode == '6'">
         <div class="row mt-3">
           <div class="col-1">
             <h2>Faq</h2>
@@ -254,10 +252,10 @@
             </tr>
           </tbody>
         </table>
-      </div>
+      </div> -->
 
       <!-- !Cart -->
-      <div *ngIf="blockCode == '7'">
+      <!-- <div *ngIf="blockCode == '7'">
         <div class="mt-3">
           <h2>Carts</h2>
         </div>
@@ -277,13 +275,13 @@
           </thead>
           <tbody *ngIf="displayData">
             <tr *ngFor="let data of displayData">
-              <td>{{data.cartId}}</td>
+              <td>{{data.cartId}}</td> -->
               <!--CartProds-->
-              <td>
+              <!-- <td>
                 <div class="dropdown">
                   <button class="btn btn-secondary" type="button" (click)="openCart(data)">
                     View Detail
-                  </button>
+                  </button> -->
                   <!-- <div class="dropdown-menu" aria-labelledby="dropdownMenuButton" *ngIf="data.cartProd">
                     <div *ngFor="let cart of data.cartProd">
                       <div class="dropdown-item">
@@ -294,11 +292,11 @@
                     </div>
                     <a class="dropdown-item" *ngIf="data.cartProd.length==0">Empty</a>
                   </div> -->
-                </div>
-              </td>
+                <!-- </div>
+              </td> -->
               <!--End CartProds-->
               <!--Contact-->
-              <td>
+              <!-- <td>
                 <div class="dropdown">
                   <button class="btn btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton"
                     data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
@@ -313,20 +311,20 @@
                     </div>
                   </div>
                 </div>
-              </td>
+              </td> -->
               <!--End Contact-->
-              <td>${{data.price}}</td>
+              <!-- <td>${{data.price}}</td>
               <td>{{data.location}}</td>
               <td style="font-size:10px;">{{data.createOn | date:'mediumDate'}}</td>
               <td style="font-size:10px;">{{data.plannedTime | date:'short'}}</td>
-              <td style="font-size:10px;">{{data.updateOn | date:'mediumDate'}}</td>
+              <td style="font-size:10px;">{{data.updateOn | date:'mediumDate'}}</td> -->
               <!-- <td><i class="far fa-edit" (click)="openCart(data)" title="Edit"></i></td> -->
-              <td ><i class="far fa-trash-alt" title="Delete" (click)="deleteCart(data)"></i></td>
+              <!-- <td ><i class="far fa-trash-alt" title="Delete" (click)="deleteCart(data)"></i></td>
             </tr>
           </tbody>
         </table>
       </div>
-    </div>
+    </div> -->
     </main>
   </div>
 </div>
