+++ src/app/fcomponents/basic/loading/loading.component.html
@@ -1,6 +1,9 @@
-<div>
+<div *ngIf="!feedbackMessage">
     Loading Data...Please wait
     <div class="spinner-border text-info ml-2" role="status">
       <span class="sr-only">Loading...</span>
     </div>
+</div>
+<div *ngIf="feedbackMessage" class="alert alert-danger" role="alert">
+  {{feedbackMessage}}
 </div>
