+++ src/environments/environment.prod.ts
@@ -3,4 +3,6 @@ export const environment = {
   baseUrl: 'http://luxedreameventhire.co.nz:5000/api',
   // baseUrl: 'http://localhost:5000/Api',
   baseLink: 'http://luxedreameventhire.co.nz:5000/'
+  // baseLink: 'http://localhost:5000/'
 };
+
